requirements.txt
fastapi==0.95.1
uvicorn==0.23.1
chromadb==0.3.7
sentence-transformers==2.2.0
PyMuPDF==1.21.2
python-docx==0.8.11
numpy==1.24.2